refactor(sessuib): session create & test
